Starfield fixes, If statement fix
The zip file for Starfield is not named the same as the inner subfolder, so did some variable munching to get that sorted.
And my nemesis has been found and fixed.  In an IF statement, -eq =/= =.  When you do "If ($var = ""), it WILL set that $var to nothing rather than compare.